makeproperties writable
- Make all `value`-type properties writable (not `current`, since it's just a getter)
- Enables mocking/stubbing/spying on event functions, among other things
- Add an example integration test

Without this change, it's impossible to know if an event function has been called (unless you wrap it in a *second* function, then spy on *that* function instead) via a spying library (such as Sinon.JS or Jasmine's `createSpy()`).  See [this REPL](https://tonicdev.com/boneskull/fsm-as-promised-writable-event-properties) for an example.